struktur data

+++ database/migrations/2025_10_23_034446_create_products_table.php
@@ -5,15 +5,18 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
-    public function up(): void {
+    public function up(): void
+    {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->foreignId('category_id')->constrained('product_categories')->onDelete('cascade');
             $table->foreignId('supplier_id')->nullable()->constrained('suppliers')->onDelete('set null');
             $table->string('name');
             $table->string('sku')->unique();
-            $table->decimal('price', 12, 2);
-            $table->decimal('cost_price', 12, 2)->default(0);
+
+            $table->unsignedBigInteger('price'); 
+            $table->unsignedBigInteger('cost_price')->default(0);
+
             $table->integer('stock')->default(0);
             $table->string('unit', 50)->default('pcs');
             $table->foreignId('created_by')->constrained('users')->onDelete('cascade');
@@ -21,7 +24,8 @@ return new class extends Migration {
         });
     }
 
-    public function down(): void {
+    public function down(): void
+    {
         Schema::dropIfExists('products');
     }
 };
