+++ database/migrations/2025_10_23_034742_create_transactions_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -14,16 +13,20 @@ return new class extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
             $table->string('invoice_number')->unique();
+
             $table->foreignId('customer_id')->nullable()->constrained('customers')->onDelete('set null');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->foreignId('payment_method_id')->constrained('payment_methods')->onDelete('cascade');
+
             $table->integer('total_qty');
-            $table->decimal('total_price', 12, 2);
-            $table->decimal('discount', 12, 2)->default(0);
-            $table->decimal('tax', 12, 2)->default(0);
-            $table->decimal('grand_total', 12, 2);
-            $table->decimal('paid_amount', 12, 2);
-            $table->decimal('change_amount', 12, 2)->default(0);
+
+            $table->unsignedBigInteger('total_price');
+            $table->unsignedBigInteger('discount')->default(0);
+            $table->unsignedBigInteger('tax')->default(0);
+            $table->unsignedBigInteger('grand_total');
+            $table->unsignedBigInteger('paid_amount');
+            $table->unsignedBigInteger('change_amount')->default(0);
+
             $table->enum('status', ['paid', 'unpaid', 'cancelled'])->default('paid');
             $table->timestamps();
         });
